pyproject.toml
[project]
name = "hackathon-01v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-groq>=0.2.4",
    "langchain-openai>=0.3.7",
    "phidata>=2.7.10",
    "pip>=25.0.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "yfinance>=0.2.54",
]
